Split off the database dumping into a separate source file.
